문서화 되어 있지 않던 기능을 문서화합니다
