+++ docs/src/pages/docs/api/convertQwertyToHangul.en.mdx
@@ -1,0 +1,38 @@
+---
+title: convertQwertyToHangul
+---
+
+import { Sandpack } from '@/components/Sandpack';
+
+# convertQwertyToHangul
+
+Converts English alphabets into Korean characters and sentences corresponding to the QWERTY keyboard layout.
+
+```typescript
+function convertQwertyToHangul(word: string): string;
+```
+
+## Examples
+
+```typescript
+convertQwertyToHangul('abc'); // '뮻'
+convertQwertyToHangul('ABC'); // '뮻'
+convertQwertyToHangul('vm론트'); // '프론트'
+convertQwertyToHangul('RㅏㄱEㅜrl'); // '깍뚜기'
+```
+
+## Demo
+
+<br />
+
+<Sandpack>
+
+```ts index.ts
+import { convertQwertyToHangul } from 'es-hangul';
+
+console.log(convertQwertyToHangul('abc'));
+console.log(convertQwertyToHangul('ABC'));
+console.log(convertQwertyToHangul('vm론트'));
+```
+
+</Sandpack>
