+++ docs/src/pages/docs/api/convertQwertyToHangul.ko.mdx
@@ -1,0 +1,38 @@
+---
+title: convertQwertyToHangul
+---
+
+import { Sandpack } from '@/components/Sandpack';
+
+# convertQwertyToHangul
+
+영어 알파벳을 qwerty 자판과 매칭과는 한글 문자와 문장으로 변환합니다.
+
+```typescript
+function convertQwertyToHangul(word: string): string;
+```
+
+## Examples
+
+```typescript
+convertQwertyToHangul('abc'); // '뮻'
+convertQwertyToHangul('ABC'); // '뮻'
+convertQwertyToHangul('vm론트'); // '프론트'
+convertQwertyToHangul('RㅏㄱEㅜrl'); // '깍뚜기'
+```
+
+## 사용해보기
+
+<br />
+
+<Sandpack>
+
+```ts index.ts
+import { convertQwertyToHangul } from 'es-hangul';
+
+console.log(convertQwertyToHangul('abc'));
+console.log(convertQwertyToHangul('ABC'));
+console.log(convertQwertyToHangul('vm론트'));
+```
+
+</Sandpack>
