+++ src/index.ts
@@ -5,7 +5,8 @@ export { canBeJongseong } from './canBeJongseong';
 export { canBeJungseong } from './canBeJungseong';
 export { combineCharacter } from './combineCharacter';
 export { combineVowels } from './combineVowels';
-export { convertQwertyToAlphabet, convertQwertyToHangul } from './convertQwertyToAlphabet';
+export { convertQwertyToAlphabet } from './convertQwertyToAlphabet';
+export { convertQwertyToHangul } from './convertQwertyToHangul';
 export { days } from './days';
 export { disassemble } from './disassemble';
 export { disassembleCompleteCharacter } from './disassembleCompleteCharacter';
